+++ TPC-Gomez-Chavero/domain/Product.cs
@@ -8,7 +8,7 @@ namespace domain
 {
     public class Product
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
         public ProductBranch Marca { get; set; }
@@ -16,6 +16,6 @@ namespace domain
         public ProductType Tipo { get; set; }
         public int Stock { get; set; }
         public int StockMinimo { get; set; }
-        public int PorcentajeVenta { get; set; }
+        public short PorcentajeVenta { get; set; }
     }
 }
